refactor: :lipstick: Update cardIcon styles and refactor data gridProjects
