+++ components/Cards/CardIcon/CardIcon.module.css
@@ -1,5 +1,5 @@
 .container{
-  padding: 5px 20px 25px 20px;
+  padding: 50px 40px 40px;
   display: flex;
   justify-content: center;
   align-items: center;
