+++ components/Cards/CardPrice/CardPrice.module.css
@@ -129,8 +129,8 @@
 /* Large devices (laptops/desktops, 992px and up) */
 @media only screen and (min-width: 992px) {
   .cardPrice{
-    margin: 30px;
-    width: 400px;
+    width: 470px;
+    padding: 50px 30px;
   }
 
   .descriptionCard{
@@ -140,8 +140,17 @@
   .price{
     font-size: 50px;
   }
+
+  .textItemList{
+    font-size: 20px;
+    line-height: 50px;
+  }
 }
 /* Extra large devices (large laptops and desktops, 1200px and up) */
-@media only screen and (min-width: 1200px) {}
+@media only screen and (min-width: 1200px) {
+  .cardPrice{
+    margin: 30px;
+  }
+}
 /* Extra large devices (large laptops and desktops, 1200px and up) */
 @media only screen and (min-width: 1500px) {}
