+++ components/GridImage/GridImage.tsx
@@ -2,11 +2,6 @@
 
 import { useState, useEffect } from "react";
 
-import portfolio1 from "@/public/portfolio-01.jpg";
-import portfolio2 from "@/public/portfolio-02.jpg";
-import portfolio3 from "@/public/portfolio-03.jpg";
-import portfolio4 from "@/public/portfolio-04.jpg";
-import portfolio5 from "@/public/portfolio-05.jpg";
 import ButtonOption from "@/components/GridImage/ButtonOption/ButtonOption";
 import styles from "@/components/GridImage/GridImage.module.css";
 import ImageProject, {
@@ -14,56 +9,13 @@ import ImageProject, {
 } from "@/components/GridImage/ImageProject/ImageProject";
 import { ScrollComponent } from "@/components/ScrollComponent/ScrollComponent";
 
-export function GridImage() {
+export function GridImage({ projects }: { projects: ParamsImageProject[] }) {
   const [optionSelect, setOptionSelect] = useState("All");
   const [projectsObjects, setProjectsObjects] = useState<JSX.Element[]>();
-
-  const projects: ParamsImageProject[] = [
-    {
-      alt: "seat",
-      categories: ["Branding", "Graphic", "All"],
-      src: portfolio1,
-      subtitle: "iosdesign",
-      title: "Creative Agency",
-    },
-    {
-      alt: "bike",
-      categories: ["Web", "Graphic", "All"],
-      src: portfolio2,
-      subtitle: "Branding",
-      title: "Rent bike",
-    },
-    {
-      alt: "laptop",
-      categories: ["Graphic", "All"],
-      src: portfolio3,
-      subtitle: "Web application",
-      title: "All Volees",
-    },
-    {
-      alt: "orange",
-      categories: ["Branding", "All"],
-      src: portfolio4,
-      subtitle: "Brandingdesign",
-      title: "Larq",
-    },
-    {
-      alt: "sofa",
-      categories: ["Branding", "Web", "All"],
-      src: portfolio5,
-      subtitle: "Branding",
-      title: "Trendy Design",
-    },
-    {
-      alt: "seat",
-      categories: ["Web", "All"],
-      src: portfolio1,
-      subtitle: "iosdesign",
-      title: "Jet Airplane",
-    },
-  ];
+  const [categories, setCategories] = useState<string[]>([]);
 
   useEffect(() => {
+    getUniqueCategories();
     const components = projects.map((project, index) => (
       <ImageProject
         key={index}
@@ -98,6 +50,14 @@ export function GridImage() {
     setProjectsObjects(components);
   };
 
+  const getUniqueCategories = () => {
+    const allCategories = projects.flatMap((item) => item.categories);
+    const uniqueCategories = Array.from(new Set(allCategories));
+    const result = uniqueCategories.sort((a, b) => a.localeCompare(b));
+
+    setCategories(result);
+  };
+
   return (
     <div className={styles.container}>
       <div className={styles.filterContainer}>
@@ -105,7 +65,15 @@ export function GridImage() {
           <h2 className={styles.title}>Some of our finest work.</h2>
         </ScrollComponent>
         <div className={styles.buttons}>
-          <ButtonOption
+          {categories.map((category, index) => (
+            <ButtonOption
+              key={index}
+              isChecked={optionSelect == category}
+              text={category}
+              onClick={filter}
+            />
+          ))}
+          {/* <ButtonOption
             isChecked={optionSelect == "All"}
             text="All"
             onClick={filter}
@@ -124,7 +92,7 @@ export function GridImage() {
             isChecked={optionSelect == "Graphic"}
             text="Graphic"
             onClick={filter}
-          />
+          /> */}
         </div>
       </div>
       <div className={styles.gridContainer}>{projectsObjects}</div>
