+++ pages/home/section10/Section10.module.css
@@ -78,10 +78,12 @@
 .imageBrandDark{
   filter: contrast(0);
   transition: all .5s;
+  height: fit-content;
 }
 
 .imageBrandLigth{
   transition: all .5s;
+  height: fit-content;
 }
 
 .containerImage:hover>.imageBrandDark, .containerImage:hover>.imageBrandLigth{
