+++ pages/home/section11/Section11.module.css
@@ -40,6 +40,7 @@
   line-height: 26px;
   font-weight: 400;
   text-align: center;
+  max-width: 700px;
 }
 
 .downContainer{
