+++ pages/home/section6/Section6.module.css
@@ -38,6 +38,7 @@
   line-height: 26px;
   font-weight: 400;
   text-align: center;
+  max-width: 700px;
 }
 
 .bottomContainer{
@@ -190,6 +191,10 @@
     grid-column-gap: 30px;
   }
 
+  .box{
+    padding: 30px;
+  }
+
   .title{
     font-size: 40px;
     line-height: 52px;
@@ -235,5 +240,9 @@
 @media only screen and (min-width: 1500px) {
   .container{
     max-width: 1250px;
+  }
+
+  .box{
+    padding: 60px;
   }
 }
