+++ pages/home/section7/Section7.module.css
@@ -43,6 +43,7 @@
   line-height: 26px;
   font-weight: 400;
   text-align: center;
+  max-width: 700px;
 }
 
 .bottomContainer{
