+++ pages/services/Development/section4/Section4.module.css
@@ -9,7 +9,7 @@
 }
 
 .secctionDark{
-  background-color: #242526;
+  background-color: var(--background-color-dark);
   color: #d3d3d4;
 }
 
@@ -17,36 +17,20 @@
   display: flex;
   flex-direction: column;
   padding: 50px 50px 50px 50px;
-  width: 100%;
 }
 
-.topContainer{
-  display: flex;
-  flex-direction: column;
-  align-items: center;
+.label{
+  background: rgba(0, 122, 255, .1);
+  color: #007aff;
 }
 
-.downContainer{
-  display: flex;
-  flex-direction: column;
-  margin-top: 50px;
-  z-index: 0;
-}
-
-.title{
-  font-size: 30px;
-  line-height: 40px;
+.button{
+  align-self: center;
+  max-width: 400px;
+  width: 90%;
+  margin-top: 20px;
+  font-size: 15px;
   font-weight: 700;
-  margin: 10px 0;
-  text-align: center;
-}
-
-.text{
-  font-size: 16px;
-  line-height: 26px;
-  font-weight: 400;
-  text-align: center;
-  max-width: 700px;
 }
 
 /* Extra small devices (phones, 600px and down) */
@@ -58,22 +42,6 @@
   .container{
     padding: 80px 100px 50px 100px;
   }
-
-  .title{
-    margin-bottom: 15px;
-    font-size: 35px;
-    line-height: 42px;
-  }
-  
-  .text{
-    font-size: 18px;
-    line-height: 28px;
-  }
-
-  .downContainer{
-    flex-direction: row;
-    justify-content: center;
-  }
 }
 /* Large devices (laptops/desktops, 992px and up) */
 @media only screen and (min-width: 992px) {
@@ -81,26 +49,11 @@
     max-width: 900px;
     padding: 120px 50px 100px 50px;
   }
-
-  .title{
-    font-size: 40px;
-    line-height: 52px;
-  }
-  
-  .text{
-    font-size: 20px;
-    line-height: 32px;
-  }
 }
 /* Extra large devices (large laptops and desktops, 1200px and up) */
 @media only screen and (min-width: 1200px) {
   .container{
     max-width: 1100px;
-  }
-
-  .title{
-    font-size: 50px;
-    line-height: 62px;
   }
 }
 /* Extra large devices (large laptops and desktops, 1200px and up) */
